fix(library): inline config from js-lib
This is an attempt at fixing issues with missing lock file maintenance.

The config:js-lib only includes the base config and what we have
replaced it with here. It might be that the base config is overriding
some of the coding we want in our own base which is added before.