fix(library): automerge lock file maintenance
For libraries, the lock files are not used by consumers, only by
the library itself. This makes it safe to automerge in respect to
the consumers.

If a library breaks by updating its lockfile (e.g. the build stops
working), consumers will also normally be broken by just using
the library.